# contracts/api-contract-draft.yaml
openapi: 3.1.0
info:
  title: Patika Beosztás API
  version: 0.3.0-phase2b
  description: >-
    A Phase 2B frontend–backend szerződése. A futó backend
    /openapi/v1.json dokumentuma a generált frontend típusok végleges forrása.
    A frontend HTTPS-en, az API-val azonos site alatt fusson, és a cookie-s
    kérésekhez credentials: include beállítást használjon.
servers:
  - url: /
...
paths:
...
  /api/admin/locations:
    get:
      tags: [Locations]
      summary: Telephelyek listázása
      x-required-permission: ManageLocations
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/IncludeInactive'
      responses:
        '200':
          description: Lapozott telephelylista
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedLocations'
        '403':
          $ref: '#/components/responses/Problem'
    post:
      tags: [Locations]
      summary: Telephely létrehozása
      x-required-permission: ManageLocations
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLocationRequest'
      responses:
        '201':
          description: Létrehozott telephely
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationResponse'
        '422':
          $ref: '#/components/responses/Problem'
  /api/admin/locations/{id}:
    parameters:
      - $ref: '#/components/parameters/EntityId'
    get:
      tags: [Locations]
      summary: Telephely lekérése
      x-required-permission: ManageLocations
      responses:
        '200':
          description: Telephely
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationResponse'
        '404':
          $ref: '#/components/responses/Problem'
    put:
      tags: [Locations]
      summary: Telephely módosítása vagy deaktiválása
      x-required-permission: ManageLocations
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLocationRequest'
      responses:
        '200':
          description: Módosított telephely
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationResponse'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
...
  /api/admin/locations/{locationId}/shift-templates:
    parameters:
      - $ref: '#/components/parameters/LocationId'
    get:
      tags: [Location shift templates]
      summary: Telephely műszaksablonjainak listázása
      x-required-permission: ManageLocations
      parameters:
        - $ref: '#/components/parameters/IncludeInactive'
      responses:
        '200':
          description: Műszaksablonok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LocationShiftTemplateResponse'
        '404': { $ref: '#/components/responses/Problem' }
    post:
      tags: [Location shift templates]
      summary: Telephelyi műszaksablon létrehozása
      x-required-permission: ManageLocations
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLocationShiftTemplateRequest'
      responses:
        '201':
          description: Létrehozott műszaksablon
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationShiftTemplateResponse'
        '404': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
  /api/admin/location-shift-templates/{id}:
    parameters:
      - $ref: '#/components/parameters/EntityId'
    put:
      tags: [Location shift templates]
      summary: Telephelyi műszaksablon módosítása
      x-required-permission: ManageLocations
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLocationShiftTemplateRequest'
      responses:
        '200':
          description: Módosított műszaksablon
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationShiftTemplateResponse'
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
  /api/admin/location-shift-templates/{id}/deactivate:
    parameters:
      - $ref: '#/components/parameters/EntityId'
    post:
      tags: [Location shift templates]
      summary: Telephelyi műszaksablon inaktiválása
      x-required-permission: ManageLocations
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VersionRequest'
      responses:
        '200':
          description: Inaktivált műszaksablon
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationShiftTemplateResponse'
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: __Host-PatikaSession
  parameters:
    CsrfToken:
      in: header
      name: X-CSRF-TOKEN
      required: true
      schema:
        type: string
    EntityId:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
...
    LocationId:
      in: path
      name: locationId
      required: true
      schema:
        type: string
        format: uuid
    Page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      in: query
      name: pageSize
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    Search:
      in: query
      name: search
      schema:
        type: [string, 'null']
    IncludeInactive:
      in: query
      name: includeInactive
      schema:
        type: boolean
        default: false
  responses:
    Problem:
      description: RFC 7807 ProblemDetails
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
  schemas:
...
    LocationType:
      type: string
      enum: [Central, Branch]
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: [string, 'null']
        title:
          type: [string, 'null']
        status:
          type: [integer, 'null']
        detail:
          type: [string, 'null']
        instance:
          type: [string, 'null']
        code:
          type: [string, 'null']
        errors:
          type: [array, 'null']
          items:
            type: object
            properties:
              code: { type: string }
              message: { type: string }
              field: { type: [string, 'null'] }
...
    CreateLocationRequest:
      type: object
      required: [name, type, isActive]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        type:
          $ref: '#/components/schemas/LocationType'
        address: { type: [string, 'null'], maxLength: 500 }
        isActive: { type: boolean }
    UpdateLocationRequest:
      allOf:
        - $ref: '#/components/schemas/CreateLocationRequest'
        - type: object
          required: [expectedVersion]
          properties:
            expectedVersion: { type: integer, minimum: 1 }
    LocationResponse:
      allOf:
        - $ref: '#/components/schemas/CreateLocationRequest'
        - type: object
          required: [id, version, createdAtUtc, updatedAtUtc]
          properties:
            id: { type: string, format: uuid }
            version: { type: integer, minimum: 1 }
            createdAtUtc: { type: string, format: date-time }
            updatedAtUtc: { type: string, format: date-time }
...
    VersionRequest:
      type: object
      additionalProperties: false
      required: [expectedVersion]
      properties:
        expectedVersion: { type: integer, minimum: 1 }
...
    DayOfWeek:
      type: string
      enum: [Sunday, Monday, Tuesday, Wednesday, Thursday, Friday, Saturday]
...
    ShiftTemplateCategory:
      type: string
      enum: [Morning, Afternoon, Long, Custom]
    StaffingCapability:
      type: string
      enum: [Pharmacist, SpecialistPharmacist, SpecialistAssistant, Assistant, Cleaner, Finance, Other]
...
    CreateLocationShiftTemplateRequest:
      type: object
      additionalProperties: false
      required: [name, category, weekdays, startTime, endTime, isActive]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        category: { $ref: '#/components/schemas/ShiftTemplateCategory' }
        weekdays:
          type: array
          minItems: 1
          uniqueItems: true
          items: { $ref: '#/components/schemas/DayOfWeek' }
        startTime: { type: string, format: time }
        endTime: { type: string, format: time }
        isActive: { type: boolean }
        requiredCapability:
          oneOf:
            - $ref: '#/components/schemas/StaffingCapability'
            - type: 'null'
    UpdateLocationShiftTemplateRequest:
      allOf:
        - $ref: '#/components/schemas/CreateLocationShiftTemplateRequest'
        - type: object
          required: [expectedVersion]
          properties:
            expectedVersion: { type: integer, minimum: 1 }
    LocationShiftTemplateResponse:
      allOf:
        - $ref: '#/components/schemas/CreateLocationShiftTemplateRequest'
        - type: object
          required: [id, locationId, locationName, version, createdAtUtc, updatedAtUtc]
          properties:
            id: { type: string, format: uuid }
            locationId: { type: string, format: uuid }
            locationName: { type: string }
            version: { type: integer, minimum: 1 }
            createdAtUtc: { type: string, format: date-time }
            updatedAtUtc: { type: string, format: date-time }
...
    PagedLocations:
      type: object
      required: [items, page, pageSize, totalCount]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/LocationResponse'
        page: { type: integer }
        pageSize: { type: integer }
        totalCount: { type: integer }
...
security:
  - cookieAuth: []